get submitDisabled():boolean {
        let enabled: boolean = this.elementform.valid;
        if (this.id === void 0) {
            enabled = enabled && (this.fileUpload.files.length > 0);
        }

        return !enabled;
    }

+++ Software/host/Controllers/HallsController.cs
@@ -36,7 +36,7 @@ namespace host.Controllers
         [HttpPost]
         public Task Post([FromBody]CreateHallCommand value)
         {
-            return CommandDispatcher.DispatchAsync<CreateHallCommand, long>(value);
+            return CommandDispatcher.DispatchAsync<CreateHallCommand, int>(value);
         }
 
         // PUT api/halls/5
